pyproject.toml
[project]
name = "handxarm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mujoco>=3.6.0",
    "pyserial>=3.5",
]
